requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
Pillow
